homepage editing fixed, admin menu added, more authentication checks added, code cleared a bit

+++ application/controllers/PageMaintenance.php
@@ -11,15 +11,21 @@ class PageMaintenance extends CI_Controller
     }
  
     function index()
-    {
-        //this function will retrive all entry in the database
+    {        
         $this->load->library('session');
+        if(!$this->session->userdata('userlogin'))
+            redirect(PRE_INDEX_URL.'index.php/');
+        
         $data['query'] = $this->pagemaintenance_model->get_all_pages();
         $this->load->view('pagemaintenance/index',$data);
     }
  
     function delete()
     {
+        $this->load->library('session');
+        if(!$this->session->userdata('userlogin'))
+            redirect(PRE_INDEX_URL.'index.php/');
+        
         $whichpage = $this->input->get('id');
         
         $this->pagemaintenance_model->delete($whichpage);        
@@ -34,7 +40,7 @@ class PageMaintenance extends CI_Controller
         $whichpage = $this->input->get('id');
         
         $data['query'] = $this->pagemaintenance_model->toUpdate($whichpage);
-        //print_r($data['query']);die;
+        
         $this->load->view('pagemaintenance/update',$data);
     }
     
@@ -44,8 +50,7 @@ class PageMaintenance extends CI_Controller
         $this->load->library(array('form_validation','session'));
         if(!$this->session->userdata('userlogin'))
             redirect(PRE_INDEX_URL.'index.php/');
- 
-        //set validation rules
+         
         $this->form_validation->set_rules('entry_name', 'Title', 'required|max_length[200]');
         $this->form_validation->set_rules('entry_body', 'Body', 'required');
  
@@ -57,6 +62,9 @@ class PageMaintenance extends CI_Controller
         else
         {
             //if valid
+            if(!$this->input->post('entry_name'))
+            { redirect(PRE_INDEX_URL.'index.php'); end;}
+            
             $name = $this->input->post('entry_name');
             $name = mb_ereg_replace("([^\w\s\d\-_~,;\[\]\(\).])", '', $name);
             $name = mb_ereg_replace("([\.]{2,})", '', $name);
